% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common_effect.R
\docType{data}
\name{DAG.p53}
\alias{DAG.p53}
\title{with an estimate of', E.DICT[\link{i}], '\n')
with an estimate of', E.DICT[\link{i}], '\n')}}}
\format{
An object of class \code{dagitty} of length 1.
}
\usage{
DAG.p53
}
\description{
with an estimate of', E.DICT[\link{i}], '\n')
with an estimate of', E.DICT[\link{i}], '\n')}}
}
\keyword{datasets}
